pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.11",
    "django-widget-tweaks>=1.5.1",
    "gunicorn>=25.1.0",
    "pillow>=12.1.1",
    "psycopg2-binary>=2.9.11",
    "pytest>=9.0.2",
    "pytest-django>=4.12.0",
    "python-dotenv>=1.2.1",
    "whitenoise>=6.11.0",
]
